requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
catboost==1.2.5
numpy==1.24.3
schedule==1.2.0
